feature: repeat() will now work with add() called from normal mode
Instead of using the g@ operator and letting the user complete the rest
of the necessary input, the user will be prompted first for the motion,
then for the surrounding character separately.

A function that validates the users motion input has been added for that
purpose.

The vimL file containing a custom function was no longer necessary. The
one function it contained itself became obsolete and the new
SurroundNoop() function will be added as part of setup() via vim.cmd().