feature: allow for arbitrary brackets via aliases
The setup function will now generate the pairs table by taking the
user's input and adding the default values to that without overwriting
them. Additionally, a new setup option (space_on_alias) was added that
determines whether the usage of the alias should add spaces within the
brackets.

The README was changed to inform about this new change.

parser.get_pair_positions() was changed to look not just at a single
character, but at a string equal to the length of the bracket.