refactor: merge all get_..._pair() functions in parser.lua
Because of the differences in implementation in nested and non-nested
pairs functions it was not possible to delete non-nested when they were
on different lines.

The same code was duplicated SIX times throughout parser.lua with only
minimal differences in variable names. They were all merged into
a single function that cater to all needs.