fix: insert mode maps work with arbitrary brackets
Only works reliably when brackets are composed of only single-byte
characters or brackets are put on separate lines. This is because lua's
pattern matching works byte-wise, not character-wise. Multi-byte
characters cause the system to insert too many <Left>s into the map
which causes the mapping to end somewhere within the closing bracket.