fix: properly find brackets when inside one
To find a closing bracket when the cursor is inside the closing bracket,
the search needs to start earlier. In that case when it finds an opening
bracket before the cursor position that does not count towards the
pair_stack.

Example:
```
      cursor
         v
   <div></div>
   ^
start of search
```
To delete the surrounding div tag when the cursor is at the marked
position, the start of the search needs to begin several characters
earlier to find the closing </div> tag. For this purpose the start of
the search is pushed back by the length of the closing bracket.
In this case it will find the opening <div> tag first. This must not
increment pair_stack, since it is not a nested pair, but the correct
counter to the selected closing tag.